Separated helper functions from zyre_tools file
