--- include/zyre_tools.h
@@ -1,18 +1,0 @@
-/**
- * Copyright (c) Santosh Thoduka, 2018
- * Licensed under GPLv3
- */
-
-#include <zyre.h>
-#include <iostream>
-#include <string>
-#include <sstream>
-#include <algorithm>
-#include <iterator>
-#include <vector>
-#include <map>
-#include <set>
-
-#include <readline/readline.h>
-#include <readline/history.h>
-#include <csignal>
